Many changes to noise code
Code that used Noise.Layered2D (or 3D, 4D, of 6D) before probably had an
issue that made it look like one-octave noise regardless of how many
octaves were used; this behavior has been moved into
Noise.InverseLayered2D (or 3D, etc.), while Layered2D has the
more-expected octave layering behavior. Turbulent2D, 3D, and so on have
also been changed to use the expected octave layering, without the old
version being kept. DetailedWorldMapDemo no longer has the blocky cloud
effect on zoom, and the technique with 4D noise for clouds that had
previously been used has returned to produce clouds (they start turned
off, press 'c' to make things cloudy). SeededNoise no longer does noise
calculations with floats; float math has potential precision issues at
high zoom, is not really any faster on desktop JDKs, and (I think) is
actually slower on GWT because float must be emulated. ThrustAltRNG uses
an update of the algorithm that has much better proven quality (with a
very high score in gjrand and more than 4TB of data without failures in
PractRand), while it appears to have the same speed.

+++ squidlib-util/src/main/java/squidpony/squidmath/MasonNoise.java
@@ -1436,7 +1436,7 @@ public class MasonNoise implements Noise.Noise2D, Noise.Noise3D, Noise.Noise4D, 
     }
     public static double noise(final float x, final float y, final float z, final int seed) {
         double n0, n1, n2, n3;
-        final float[] gradient3DLUT = SeededNoise.gradient3DLUT;
+        final double[] gradient3DLUT = SeededNoise.gradient3DLUT;
         final float s = (x + y + z) * F3;
         final int i = fastFloor(x + s),
                 j = fastFloor(y + s),
